web/requirements.txt
requests==2.19.1
scipy==1.1.0
numpy==1.14.6
Cython==0.28.5
pandas==0.23.4
flask==1.0.2
psycopg2-binary==2.7.5
Flask-SQLAlchemy==2.3
Flask-Migrate==2.2.1
alembic==0.8.5
flask-marshmallow==0.9.0
apscheduler==3.5.2
redis==2.10.6
backoff==1.6.0
marshmallow-sqlalchemy==0.14.1
scikit-surprise==1.0.6